common/gather_youtube/requirements.txt
yt-dlp==2024.10.7
pymongo==4.3.3
python-dotenv==0.21.1
